+++ example/app/admin.py
@@ -13,13 +13,14 @@ admin.site.has_permission = disable_admin_login()
 
 @admin.register(Blog)
 class BlogAdmin(ActiveLanguageMixin, admin.ModelAdmin):
-    pass
+    list_display = ("title_i18n", "category")
+    search_fields = ("title_i18n", "category__name_i18n")
 
 
 class DefaultAdminCategory(Category):
     """
-    Proxy model to have both the unlimited version of the ModelAdmin and
-    a limited version
+    Proxy model to have both the unlimited version of the ModelAdmin and a limited version with all
+    fields but the default language and the current language excluded.
     """
 
     class Meta: