+++ tests/test_utils.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.db import models
 from django.test import TestCase
 
 from modeltrans.manager import transform_translatable_fields